Add files via upload
Here, the login of the user is verified using the user name and the password. Then the user is taken to the transaction console where he/she can opt for various operations like deposit, withdraw, and transfer money to another bank account. User also has option to logout anytime if he/she opts so.

	A random OTP gets generated which must be confirmed by the user in order to make the transfer successful. All the choices are fully operational and displays the remaining balance.